Восстановлена компилируемость проектов Visual Studio

+++ sources/Device53/src/Ethernet/TcpSocket.cpp
@@ -131,9 +131,9 @@ static void SendAnswer(void *_arg, struct tcp_pcb *_tpcb)
         "<allow-access-from domain=\"*\" to-ports=\"9999\" />"                                                 \
         "</cross-domain-policy>"                                                                            \
         "\0";
-    struct pbuf *tcpBuffer = pbuf_alloc(PBUF_RAW, strlen(policy), PBUF_POOL);
+    struct pbuf *tcpBuffer = pbuf_alloc(PBUF_RAW, (uint16)strlen(policy), PBUF_POOL);
     tcpBuffer->flags = 1;
-    pbuf_take(tcpBuffer, policy, strlen(policy));
+    pbuf_take(tcpBuffer, policy, (uint16)strlen(policy));
     struct State *s = (struct State *)_arg;
     s->p = tcpBuffer;
     SendPCB(_tpcb, s);
